+++ CaveStoryModdingFramework/Lists/Bosses.cs
@@ -5,8 +5,8 @@ namespace CaveStoryModdingFramework
 {
     public static class Bosses
     {
-        public static readonly ReadOnlyDictionary<int, string> BossNameList
-            = new ReadOnlyDictionary<int, string>(new Dictionary<int, string>()
+        public static readonly ReadOnlyDictionary<long, string> BossNameList
+            = new ReadOnlyDictionary<long, string>(new Dictionary<long, string>()
         {
             {0, BossNames.None },
             {1, BossNames.Omega },
